resolver: simplify TLS setup
