+++ crates/resolver/Cargo.toml
@@ -28,12 +28,11 @@ license.workspace = true
 default = ["system-config", "tokio-runtime"]
 backtrace = ["dep:backtrace", "hickory-proto/backtrace"]
 dns-over-rustls = [
-    "dns-over-tls",
     "dep:rustls",
     "dep:tokio-rustls",
     "hickory-proto/dns-over-rustls",
+    "tokio-runtime",
 ]
-dns-over-tls = ["tokio-runtime"]
 
 # This requires some TLS library, currently only rustls is supported
 dns-over-https-rustls = [
@@ -123,5 +122,5 @@ workspace = true
 
 [package.metadata.cargo-all-features]
 skip_optional_dependencies = true
-denylist = ["dnssec", "dns-over-tls"]
+denylist = ["dnssec"]
 max_combination_size = 2
